Junior Web Dev Question
While reading forums for interview questions, I saw this one mentioned. Similar to Fizzbuzz, if a number is divisible by 2 it prints "a", if it is divisible by 3 it prints "b", and if it is divisible by both it prints "ab". Otherwise it just prints the number.

This game runs from 1 to 100.